Tons of refactoring, shift to slow down
Added a point constructor that holds x and y values, which led to tons of refactoring. Everything that had a separate x and y property now holds them inside of a point (except for velocities). What's more, is most objects that exist and can be drawn should have a getX, getY, changeX, changeY, setX, and setY.